feat: test setup
This includes the setup to run the tests on a k8s cluster pointing to
existing KubeCF and Minibroker installations.

The current script logs in as a CF admin, creates an org, a space, and
enables the service broker.

The logic for each provider is not included in this commmit.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>